Cleaned up macOS build
- upgraded Xcode project
- set C++14 standard
- increased deployment target to 10.9.
- added missing ROM resources to app bundle.

Also fixed many build warnings:
- abuse of the comma operator
- unused variables
- narrowing type conversions

+++ Base/PNG.h
@@ -65,7 +65,7 @@ typedef struct
 {
     DWORD dwWidth, dwHeight;
     BYTE *pbImage;
-    UINT uSize, uCompressedSize;
+    ULONG uSize, uCompressedSize;
 } PNG_INFO;
